Added controller for userExists, verifyToken and signup
